app/services/user_service.py
# services/user.py
import jwt
from datetime import datetime, timedelta, timezone
from sqlalchemy.orm import Session
from app.models import User
from app.schemas.user_schema import UserCreate


def create_user(db: Session, user_create: UserCreate):
    db_user = User(
        username=user_create.username,
        email=user_create.email,
        password=user_create.password,
    )
    db.add(db_user)
    db.commit()
    db.refresh(db_user)
    return db_user


def get_user_by_username(db: Session, username: str) -> User:
    return db.query(User).filter(User.username == username).first()


def get_user_by_email(db: Session, email: str):
    return db.query(User).filter(User.email == email).first()
